+++ Code/Engine/Foundation/Basics/Types.h
@@ -56,6 +56,7 @@ public:
   bool operator!= (ezResultEnum cmp) const { return e != cmp; }
 
   bool IsSuccess() const { return e == EZ_SUCCESS; }
+  bool IsFailure() const { return e == EZ_FAILURE; }
 
 private:
   ezResultEnum e;